# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * StockSense AI retail intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  CRITICAL: 'CRITICAL',
  LOW: 'LOW',
  HEALTHY: 'HEALTHY',
  OVERSTOCK: 'OVERSTOCK',
  DECLINING: 'DECLINING',
} as const;

export interface Product {
  id: string;
  name: string;
  category: string;
  price: number;
  currentStock: number;
  dailySales: number;
  supplierLeadTime: number;
  daysRemaining: number;
  status: ProductStatus;
  recommendation: string;
  growthPercent?: number;
  units30d?: number;
}

export interface SeriesPoint {
  label: string;
  value: number;
}

export interface HealthSlice {
  label: string;
  value: number;
  color: string;
}

export type AttentionItemSeverity = typeof AttentionItemSeverity[keyof typeof AttentionItemSeverity];


export const AttentionItemSeverity = {
  CRITICAL: 'CRITICAL',
  WARNING: 'WARNING',
  INFO: 'INFO',
} as const;

export interface AttentionItem {
  id: string;
  productId: string;
  title: string;
  description: string;
  severity: AttentionItemSeverity;
  action: string;
}

export interface Dashboard {
  totalSales: number;
  totalInventory: number;
  lowStockProducts: number;
  overstockProducts: number;
  salesGrowth: number;
  healthScore: number;
  salesSeries: SeriesPoint[];
  inventoryHealth: HealthSlice[];
  topProducts: Product[];
  slowProducts: Product[];
  attention: AttentionItem[];
}

export interface InventoryPage {
  items: Product[];
  total: number;
  categories: string[];
}

export interface Evidence {
  source: string;
  period: string;
  product: string;
  unitsSold: number;
  averageDailySales: number;
  currentStock: number;
  supplierLeadTime: number;
  daysRemaining: number;
  calculation: string;
  conclusion: string;
}

export type ProductDetail = Product & {
  previousPeriodUnits: number;
  trend: SeriesPoint[];
  evidence: Evidence;
};

export interface CategoryValue {
  category: string;
  value: number;
}

export interface Analytics {
  period: number;
  revenue: number;
  unitsSold: number;
  growthPercent: number;
  salesSeries: SeriesPoint[];
  categoryRevenue: CategoryValue[];
  topProducts: Product[];
  bottomProducts: Product[];
}

export type RecommendationPriority = typeof RecommendationPriority[keyof typeof RecommendationPriority];


export const RecommendationPriority = {
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  LOW: 'LOW',
} as const;

export interface Fact {
  label: string;
  value: string;
}

export interface Recommendation {
  id: string;
  productId: string;
  productName: string;
  priority: RecommendationPriority;
  problem: string;
  evidence: string;
  action: string;
  reorderQuantity: number;
  calculation: string;
  facts?: Fact[];
}

export interface RecommendationsResponse {
  immediate: Recommendation[];
  monitor: Recommendation[];
  noAction: Recommendation[];
}

export interface CopilotInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  question: string;
}

export interface CopilotResponse {
  question: string;
  summary: string;
  answer: string;
  evidence: Evidence[];
  assumptions: string[];
  sufficientData: boolean;
  referencedProducts?: Product[];
}

export type NotificationSeverity = typeof NotificationSeverity[keyof typeof NotificationSeverity];


export const NotificationSeverity = {
  CRITICAL: 'CRITICAL',
  WARNING: 'WARNING',
  INFO: 'INFO',
} as const;

export interface Notification {
  id: string;
  title: string;
  description: string;
  severity: NotificationSeverity;
  productId: string;
  createdAt: string;
}

export type InventoryUploadInputRowsItem = { [key: string]: unknown };

export interface InventoryUploadInput {
  rows: InventoryUploadInputRowsItem[];
}

export type SalesUploadInputRowsItem = { [key: string]: unknown };

export interface SalesUploadInput {
  rows: SalesUploadInputRowsItem[];
}

export interface UploadResult {
  imported: number;
  errors: string[];
  message: string;
}

export type ListInventoryParams = {
search?: string;
category?: string;
status?: ListInventoryStatus;
sort?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
pageSize?: number;
};

export type ListInventoryStatus = typeof ListInventoryStatus[keyof typeof ListInventoryStatus];


export const ListInventoryStatus = {
  CRITICAL: 'CRITICAL',
  LOW: 'LOW',
  HEALTHY: 'HEALTHY',
  OVERSTOCK: 'OVERSTOCK',
  DECLINING: 'DECLINING',
} as const;

export type GetAnalyticsParams = {
period?: GetAnalyticsPeriod;
};

export type GetAnalyticsPeriod = typeof GetAnalyticsPeriod[keyof typeof GetAnalyticsPeriod];


export const GetAnalyticsPeriod = {
  NUMBER_7: 7,
  NUMBER_30: 30,
  NUMBER_90: 90,
} as const;