# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StockSense AI retail intelligence API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: dashboard
  - name: inventory
  - name: analytics
  - name: copilot
  - name: uploads
  - name: recommendations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get executive dashboard data
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: Search and filter inventory
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [CRITICAL, LOW, HEALTHY, OVERSTOCK, DECLINING]
        - name: sort
          in: query
          schema: { type: string, default: risk }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryPage"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [inventory]
      summary: Get product detail and evidence
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Product not found
  /analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      summary: Get sales analytics
      parameters:
        - name: period
          in: query
          schema:
            type: integer
            enum: [7, 30, 90]
            default: 30
      responses:
        "200":
          description: Sales analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analytics"
  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: Get calculated recommendations
      responses:
        "200":
          description: Recommendations grouped by priority
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsResponse"
  /notifications:
    get:
      operationId: listNotifications
      tags: [dashboard]
      summary: Get current attention notifications
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /copilot:
    post:
      operationId: askCopilot
      tags: [copilot]
      summary: Ask the evidence-backed retail copilot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopilotInput"
      responses:
        "200":
          description: Copilot answer with calculated facts and evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopilotResponse"
  /demo/load:
    post:
      operationId: loadDemoData
      tags: [uploads]
      summary: Load the demo retail dataset
      responses:
        "200":
          description: Demo data loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
  /upload/inventory:
    post:
      operationId: uploadInventory
      tags: [uploads]
      summary: Validate and import inventory CSV rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryUploadInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
  /upload/sales:
    post:
      operationId: uploadSales
      tags: [uploads]
      summary: Validate and import sales CSV rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalesUploadInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
components:
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema: { type: string }
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]
    Product:
      type: object
      required: [id, name, category, price, currentStock, dailySales, supplierLeadTime, daysRemaining, status, recommendation]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        price: { type: number }
        currentStock: { type: integer }
        dailySales: { type: number }
        supplierLeadTime: { type: integer }
        daysRemaining: { type: number }
        status:
          type: string
          enum: [CRITICAL, LOW, HEALTHY, OVERSTOCK, DECLINING]
        recommendation: { type: string }
        growthPercent: { type: number }
        units30d: { type: integer }
    Dashboard:
      type: object
      required: [totalSales, totalInventory, lowStockProducts, overstockProducts, salesGrowth, healthScore, salesSeries, inventoryHealth, topProducts, slowProducts, attention]
      properties:
        totalSales: { type: number }
        totalInventory: { type: integer }
        lowStockProducts: { type: integer }
        overstockProducts: { type: integer }
        salesGrowth: { type: number }
        healthScore: { type: integer }
        salesSeries:
          type: array
          items: { $ref: "#/components/schemas/SeriesPoint" }
        inventoryHealth:
          type: array
          items: { $ref: "#/components/schemas/HealthSlice" }
        topProducts:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        slowProducts:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        attention:
          type: array
          items: { $ref: "#/components/schemas/AttentionItem" }
    InventoryPage:
      type: object
      required: [items, total, categories]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        total: { type: integer }
        categories:
          type: array
          items: { type: string }
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          required: [previousPeriodUnits, evidence, trend]
          properties:
            previousPeriodUnits: { type: integer }
            trend:
              type: array
              items: { $ref: "#/components/schemas/SeriesPoint" }
            evidence: { $ref: "#/components/schemas/Evidence" }
    Analytics:
      type: object
      required: [period, revenue, unitsSold, growthPercent, salesSeries, categoryRevenue, topProducts, bottomProducts]
      properties:
        period: { type: integer }
        revenue: { type: number }
        unitsSold: { type: integer }
        growthPercent: { type: number }
        salesSeries:
          type: array
          items: { $ref: "#/components/schemas/SeriesPoint" }
        categoryRevenue:
          type: array
          items: { $ref: "#/components/schemas/CategoryValue" }
        topProducts:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        bottomProducts:
          type: array
          items: { $ref: "#/components/schemas/Product" }
    RecommendationsResponse:
      type: object
      required: [immediate, monitor, noAction]
      properties:
        immediate:
          type: array
          items: { $ref: "#/components/schemas/Recommendation" }
        monitor:
          type: array
          items: { $ref: "#/components/schemas/Recommendation" }
        noAction:
          type: array
          items: { $ref: "#/components/schemas/Recommendation" }
    Recommendation:
      type: object
      required: [id, productId, productName, priority, problem, evidence, action, reorderQuantity, calculation]
      properties:
        id: { type: string }
        productId: { type: string }
        productName: { type: string }
        priority:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        problem: { type: string }
        evidence: { type: string }
        action: { type: string }
        reorderQuantity: { type: integer }
        calculation: { type: string }
        facts:
          type: array
          items: { $ref: "#/components/schemas/Fact" }
    CopilotInput:
      type: object
      required: [question]
      properties:
        question: { type: string, minLength: 1, maxLength: 500 }
    CopilotResponse:
      type: object
      required: [question, summary, answer, evidence, assumptions, sufficientData]
      properties:
        question: { type: string }
        summary: { type: string }
        answer: { type: string }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/Evidence" }
        assumptions:
          type: array
          items: { type: string }
        sufficientData: { type: boolean }
        referencedProducts:
          type: array
          items: { $ref: "#/components/schemas/Product" }
    Evidence:
      type: object
      required: [source, period, product, unitsSold, averageDailySales, currentStock, supplierLeadTime, daysRemaining, calculation, conclusion]
      properties:
        source: { type: string }
        period: { type: string }
        product: { type: string }
        unitsSold: { type: integer }
        averageDailySales: { type: number }
        currentStock: { type: integer }
        supplierLeadTime: { type: integer }
        daysRemaining: { type: number }
        calculation: { type: string }
        conclusion: { type: string }
    Fact:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
    SeriesPoint:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: number }
    HealthSlice:
      type: object
      required: [label, value, color]
      properties:
        label: { type: string }
        value: { type: integer }
        color: { type: string }
    CategoryValue:
      type: object
      required: [category, value]
      properties:
        category: { type: string }
        value: { type: number }
    AttentionItem:
      type: object
      required: [id, productId, title, description, severity, action]
      properties:
        id: { type: string }
        productId: { type: string }
        title: { type: string }
        description: { type: string }
        severity:
          type: string
          enum: [CRITICAL, WARNING, INFO]
        action: { type: string }
    Notification:
      type: object
      required: [id, title, description, severity, productId, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        severity:
          type: string
          enum: [CRITICAL, WARNING, INFO]
        productId: { type: string }
        createdAt: { type: string }
    InventoryUploadInput:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items: { type: object, additionalProperties: true }
    SalesUploadInput:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items: { type: object, additionalProperties: true }
    UploadResult:
      type: object
      required: [imported, errors, message]
      properties:
        imported: { type: integer }
        errors:
          type: array
          items: { type: string }
        message: { type: string }